Move stats helper modules into `lib`
